+++ NEWS.md
@@ -5,6 +5,7 @@
 * Renamed the package from `@depressed-pho/cicada-build` to
   `cicada-build`. Since it's not meant to be published on the public NPM
   repository, it doesn't make sense to have a scope.
+* Relative imports in TypeScript modules now require extensions.
 
 ## 1.0.0 -- 2023-01-01